[system-mac os] Fix memory metrics
We use the output of ps for the mem usage stats, but the values can
have different units (can be G or M).

We would need to parse the unit and convert the values to the same unit
(probably bytes), otherwise it doesn't make much sense.

The Mac Package is now bundling psutil so let’s use psutil to get these
metrics (and add more!)